requirements.txt
certifi==2021.10.8
dnspython==1.16.0
email-validator==1.1.2
Flask==2.0.3
Flask_Login==0.5.0
Flask_Moment==1.0.2
flask_mongoengine==0.9.5
Flask_WTF==1.0.0
google_api_python_client==2.50.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
gunicorn==20.0.0
Jinja2==3.0.3
mail==2.1.0
mongoengine==0.20.0
matplotlib
oauthlib==3.2.0
protobuf==4.21.1
PyJWT==2.6.0
requests==2.22.0
setuptools==65.5.0
Werkzeug==2.0.3
WTForms==2.3.3
WTForms_Components==0.10.4